Remove inline symbol search and banner features
