--- client/browser/src/shared/components/symbols.scss
@@ -1,59 +1,0 @@
-.symbols-dropdown-container {
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 30;
-    transform: translateZ(0);
-}
-
-.symbols-dropdown {
-    position: relative;
-    top: 0;
-    left: 0;
-    min-width: 180px;
-    margin-top: 20px;
-    cursor: pointer;
-    background: #ffffff;
-    border: 1px solid #dfe2e5;
-    border-radius: 3px;
-    box-shadow: 0 0 5px rgba(27, 31, 35, 0.1);
-
-    &__suggestions {
-        padding: 0;
-        margin: 0;
-        list-style: none;
-        display: block;
-
-        max-height: 15rem;
-        overflow-y: auto;
-    }
-
-    &__suggestion {
-        display: flex;
-        align-items: center;
-        padding: 5px 10px;
-        font-weight: 600;
-        border-bottom: 1px solid #dfe2e5;
-
-        &--focused {
-            color: #ffffff;
-            text-decoration: none;
-            background: #0366d6;
-        }
-    }
-
-    &__suggestion:last-child {
-        border-bottom: 0;
-        border-bottom-right-radius: 3px;
-        border-bottom-left-radius: 3px;
-    }
-
-    &__filepath {
-        font-weight: 400;
-        color: #586069;
-
-        &--focused {
-            color: #ffffff;
-        }
-    }
-}
