+++ client/browser/src/shared/util/context.tsx
@@ -10,8 +10,6 @@ export let eventLogger = new EventLogger()
 export let sourcegraphUrl =
     window.localStorage.getItem('SOURCEGRAPH_URL') || window.SOURCEGRAPH_URL || DEFAULT_SOURCEGRAPH_URL
 
-export let inlineSymbolSearchEnabled = false
-
 interface UrlCache {
     [key: string]: string
 }
@@ -21,7 +19,6 @@ export const repoUrlCache: UrlCache = {}
 if (window.SG_ENV === 'EXTENSION') {
     storage.getSync(items => {
         sourcegraphUrl = items.sourcegraphURL
-        inlineSymbolSearchEnabled = items.featureFlags.inlineSymbolSearchEnabled
     })
 }
 
@@ -31,10 +28,6 @@ export function setSourcegraphUrl(url: string): void {
 
 export function isSourcegraphDotCom(url: string = sourcegraphUrl): boolean {
     return url === DEFAULT_SOURCEGRAPH_URL
-}
-
-export function setInlineSymbolSearchEnabled(enabled: boolean): void {
-    inlineSymbolSearchEnabled = enabled
 }
 
 export function getPlatformName(): 'phabricator-integration' | 'firefox-extension' | 'chrome-extension' {
